Add a very basic read-only API based on django-rest-framework
This API is still awkward to use in its current state.  The main
deficiencies that should be fixed before version 1 are:

 * The IDs for posts, organizations and elections that are used
   in the API should be their slug attributes (on the PostExtra,
   OrganizationExtra and Election models respectively).

 * There should be a fuzzy lookup of people by names (or at least
   substring lookup)

And some nice-to-have features:

 * Add filtering of results on more fields.

 * It would be good to have the equivalent of PopIt's embed parameter,
   so you can embed memberships, people or organization information to
   the depth you want. (At the moment it's somewhat inconsistent
   which relations have their data embedded or not.)

This also fixes the 'Open data JSON API' links on person pages to
link to this new API.  Thanks to Andy Lulham (@andylolz) for pointing
out that these links should be absolute for more useful
copy-and-pasting.