Fix naive datetime warning on running migrations from scratch
At an early stage in the development of the QueuedImage model
we had both default=datetime.datetime.now and auto(_now)_add=True for
the created/updated fields.  Using the naive datetime.datetime.now
produces a warning when timezone support is enabled (USE_TZ=True)
but even though we fixed this in a later migration to *only* use
auto(_now)_add we still get these warnings when migrating from
scratch (e.g. when setting up a new project):

    Applying
  moderation_queue.0012_auto_20150717_0811.../home/mark/.virtualenvs/ynr-popolo-upstream/local/lib/python2.7/site-packages/django/db/models/fields/__init__.py:1474:
  RuntimeWarning: DateTimeField QueuedImage.created received a naive
  datetime (2015-12-06 10:38:35.860379) while time zone support is active.
    RuntimeWarning)

  /home/mark/.virtualenvs/ynr-popolo-upstream/local/lib/python2.7/site-packages/django/db/models/fields/__init__.py:1474:
  RuntimeWarning: DateTimeField QueuedImage.updated received a naive
  datetime (2015-12-06 10:38:35.962306) while time zone support is active.
    RuntimeWarning)

Editing the early migration to remove default=datetime.datetime.now
should be safe because all current installations have migrated past
that point and the migration that removed it.

The useful post that pointed me towards the issue being with a
previous migration is:

  https://groups.google.com/d/msg/django-users/7Uzx6MFplh4/wy1W-BVCEgkJ

Fixes #484