Change the 'Posts' links to be called 'Elections'
This change is at the request of Democracy Club, as discussed in Slack:

The link used to be "Constituencies" but was changed to the more
generic "Posts" to cover multiple elections.  However, some people assume
that "Posts" means "Blog Posts".  "Areas" was suggested as an alternative,
but since there are also area pages, this doesn't seem quite right.

"Elections" seems like a resonable replacement, since it does list every
current election and its posts.