Remove the artifical membership start and end dates
Fixes #574 - to quote (slightly edited) from that issue:

When using PopIt as the data store, we previously set artificial start
and end dates on memberships representing candidacies and their parties.
(These typically started the date after the previous election and ended
on the day of the election.) This was somewhat useful so that 'at' queries
with the current date and time in PopIt worked.

This has a few problems - it adds confusing fields to the Election
object which we shouldn't need (the default values of these start / end
dates), and misleads people into thinking we know the actual date the
person joined a party or declared their candidacy, when we don't.

Now we have the election property on MembershipExtra these aren't
needed, and the current code doesn't even set them on new memberships.

So, this commit removes these fields from the Election object:

  candidacy_start_date
  party_membership_start_date
  party_membership_end_date

... and any reference to them.