Change UK postcode lookup to return all relevant elections
In the 2015 general election, there was only one election we were
handling, so after putting in a postcode we could redirect straight
to the post page.  However, in the 2016 UK elections there may be
multiple elections or election sub-types for a given postcode.

So, this commit changes the postcode lookup to be more like the
generic address lookup in YNR, which redirects to an areas page
with all the posts relevant to that location.