move get_area_ids out to election specific
As the get_area_ids call is likely to change a bit depending on the
jurisdiction define it in election specific rather than having it in the
API code.

There is a default version that just uses the co-ordinate lookup and
this has a uk version that does postcodes as well.