A command to fix constituency / region data in existing UK 2016 sites
The script used to create the elections, posts, areas, etc. for the UK
2016 regional assembly elections unfortunately mixed up the mapit_code
and the slug / name / party_lists_in_use for Scotland and Wales.

This script (part of the fix for #600) fixes the election metadata for
these regions; we change the slug / name /party_lists_in_use so that
we don't have to reassociate all the posts with a different election.