Script to generate parties and party sets for Costa Rica
Uses the parties.csv file in elections/cr/data/ to generate the parties
and party sets for the 2016 elections.

It generates a party set for each Canton from MapIt and then uses the
list of Cantons the party is standing in to add it to the appropriate
party sets.

It also adds the "Cédula Jurídica" ID as an identifier and uses this as
the slug where available, otherwise it falls back to the slugified name