Basic Haystack based search for people
Add in a custom template to make styling it easier and add an empty
search view so we can control the URL structure.

We're indexing all the names irrespective of whether they appear in the
UI as they might have been added in an import, and if we do add them to
the UI in future this way we won't forget to add them to the search
index.

ElasticSearch is set up to update the index in real time when Person
models are saved. We might want to revisit this in future and implement
a queue if it has adverse load implications.