Make DOB parsing more robust
Make sure it handles month names in the current language. This happens
with a custom parserinfo class that adds translations of all the month
names in the current language in. I am not sure this is the cleanest way
to do this.

We don't use fuzzy matching as it turns out this is *way too* fuzzy and
generates as many false positives as it does real ones. e.g.
"this is not a date" is parsed as today's date :(

This uses lazy gettext on month translations. This is partly so we can
test them but also to make sure that they are only translated when used
so we get the language at the time and not at module load time.