New header layout and candidate search box
Part of #644. The addition of a search box necessitated a reworking
of the header layout. There is now a "masthead" bar with the site
logo and user links in, and below it, a "nav" with site links and
the search box.

Some of the repeated code for the search box background image and
placeholder pseudo-element is spun out into two mixins, which live
in the new mixins.scss file.