use SimplePopoloFields to work out what to display on a person page
Becuase quite a few of the things displayed on this page are actually a
compound of several fields or use complicated formatting we're not just
looping over the list but instead checking if a field should be
displayed and then using the existing formatting to display it. This is
a bit less flexible but probably good enough and considerably less
complex than the alternative.