Create notifications when a Person is edited/created
Attach a singnal to the LoggedAction that uses django-notifications to
create events that we can send alerts for. Use LoggedAction as we only
created one of these per relevant event, whereas it's possible we save
the person several times which would create multiple alerts.

We use the LoggedAction type to filter which actions we create
notifications for - only really person editing.

Creates one event per alert for each save of a person. We need to keep
track of who we have created a notification for as we do this as it's
possible someone could have an alert for a person and the area that
person is in and we only want to create one notification in that case.

As it's the details of the person that are changing we always set the
target of the notification to be the person as it's the changes there we
will be showing