Remove post_id_format from the Election model
The post_id_format should now only ever be needed when importing new
data, so it's not necessary to have it stored in the Election model,
which already has a confusing array of attributes.

Remove the post_id_format attribute and all references to it.  (e.g.
there was still a reference to it in
candidates.static_data:BaseAreaPostData, but that whole module should
have been removed already - it's not longer used.)