Allow multiple winners to be set for a post
Previously you could only set one person as "elected" for a particular
post, but some elections in proportional representation systems allow
multiple people to be elected.

This change allows you to set more than one person as elected up to a
maximum set on the election object in the 'people_elected_per_post'
attribute. (You can also have arbitrary numbers of winners if that
field is set to -1.)