Person page: don't show a big list of elections they're not standing in
If there are 2 or fewer elections known to the site, it's probably
fine to show whether they're standing in both.  However, when there
are more elections, this begins to look weird - in the UK elections
in 2016, there'll just be a list of 10 or so "don't know" indications
in the "Constituencies" section.

So, this commit makes that cut-off, and if there are more than 2
elections known to the site, just shows the constituencies someone's
known to be standing in (or did stand in).

This also changes the section heading to "Candidacies" which is more
general than "Constituencies", and hopefully just as clear.