Model classes for complex fields
Add a model class to represent fields where the information is stored in
a generic relation such as links. This allows the customisation of the
set of such fields used in a YNR instance.

The model stores the name of the underlying relation, some details about
how it should be displayed ( label and field type ) and the details of
how to store the information in the generic relation.

The info_type_* properties are used to describe the key used to pull the
field value out of the underlying generic relation. _key being the name
of the field to store the value in info_type.

info_value_key is the name of the field in the underlying relation in
which to store the value of the complex field.