Make the areas view tolerant of some non-existent areas
You might reasonably query MapIt for areas of some particular set of
types, and form an areas URL from their types and IDs.  If only some
areas of a particular type are holding elections, and only those ones
have had Popolo areas created in YNR, you'd get an error from this
view even though some of them could be found.

This commit changes the view so it ignores areas that couldn't be
found, unless none could be found in which case it returns a 404.