Allow dynamic addition of elections on UpdatePersonForm
... and also only show the existing elections on the form by
default.

In the DemocracyClub fork this was done in:

  e3fb4df Allow dynamic addition of elections on UpdatePersonForm

This version is slightly different. In order to preserve existing
behaviour (particularly to ensure that a couple of it tests [1] still
pass) and also pick up any dynamically added candidacies, we add
extra fields for any additional elections indicated by the presence of a
standing_ field, rather than depending on an extra_election_id field.

[1] These were:

* candidates.tests.test_validators:TestValidators.test_update_person_form_standing_no_party_but_gb_constituency
* candidates.tests.test_validators:TestValidators.test_update_person_form_standing_no_party_no_constituency