Preserve the order of fields SingleElectionForm
The fields attribute from Form which was being overwritten is
an OrderedDict, and this is needed so that the question about
whether someone is standing or not in that election remains
the first input in the form.  So, overwrite self.fields with
an OrderedDict rather than a dict to make sure it's always
first.