Remove in-page JS which couldn't reference obfuscated function names
The in-page Javascript code worked OK with DEBUG = True, but in
production the person_form.js Javascript is wrapped in an IIFE
(immediately-invoked function expression) by django-pipeline and
the Javascript compressor is free to change the names of functions
defined within it, so the in-page Javascript can't find those
names any more.

There's no reason for the Javascript to be in-page, anyway, so this
commit moves it into person_form.js where references to those
functions will be obfuscated along with the definition.