Fix the GeoLocatorView for multiple MapIt generations
The logic here was wrong, so it would think there were no results
at all if the first query to MapIt returned some, but the second query
returned none.

Combined with errors in the MapIt generation data in the DC install
of YNR this meant that any geolocation query would say "Your location
does not seem to be covered by this site".

Partial fix for #751