UK: Fix the updating of existing images from the Electoral Commission
The call to ImageExtra.objects.update_or_create_from_file would find
multiple images with the same md5sum - sometimes the same image would be
used for multiple parties (e.g. when they're renamed and we still have
the old party in the database, I think).

This commit changes the criteria for selecting the image to update so
that it makes sure that it only finds images actually for the party
that's being updated.