Put the "most popular" parties at the top of the party drop-down
It was annoying that even after typing "labo" into the Select2 party
selection widget, you'd have to scroll through lots of other parties to
get the the most common choice, "Labour Party".

This commit will order the party drop-down by the number of existing
candidates, and show the number of existing candidates beside those
that have any (in order to indicate how it's ordered).

Thanks to Zarino Zappia (@zarino) for this suggestion.

Fixes #750