Sort candidates in non-party-list posts like the SOPN
The ordering of candidates for a post was confusingly different from UK
Statements of Persons Nominated where (a) party lists were not in use
and (b) there was more than one candidate from a party for that post.
(Candidates were grouped by party and then the parties were ordered by
the name of the first candidate in each party (!))

There's a huge advantage in just ordering candidates by their last name
in this case, because checking our candidates against the SOPN is way
faster and more reliable when you can step down the two lists
candidate-by-candidate checking that they correspond.  This commit
changes the ordering of candidates in the non-party list case so that
that's the case.