Clear 'not_standing' in the election when bulk adding
If a candidate is marked as not standing from the candidate
list the election is added to the not_standing relationship of
PersonExtra.  If someone is subsequently marked as standing for
any post in that election, the election needs to be removed
from not_standing.

(This is done in update_person_from_form, but that bit of the
logic was missed in BulkAddReviewView.update_person.)

Thanks to Janet McKnight (@janetmck) for finding and reporting
this problem.