Add a script to fix the not_standing relationships of people
There was a bug in bulk adding people which meant that their
"not_standing" status for an election wasn't removed when reinstating
them as a candidate in that election.

That bug has been fixed in the parent commit, but there are still people
in the database who have a candidacy (represented by a Membership and
MembershipExtra) in an election, but also have that election in their
not_standing.

This commit introduces a script that will find those cases, report them
and (if --delete is specified) fix the inconsistency by removing the
election from their not_standing.