Fix searches featuring apostrophes
The search uses the text field in the index which is generated by
Haystack and escapes the content as it goes in. Search terms were not
being escaped which meant that names with an apostrophe in them never
matched.

This escapes search terms so they should not match the values in the
text field in the search index.

Fixes #862