Fix version diffs where JSON patch operations are dependent
The operations that make up a JSON patch may depend on earlier
operations - they're meant to be applied in order.  However, the code
that parsed these to generate readable diffs assumed that each operation
would apply to the original JSON.

This commit changes that behaviour so that after dealing with each
operation it's applied to the original JSON.  This fixes the problem,
for example, where an array of three items are reordered and the
generated operations in the patch are (a) insert an item at the front
and (b) remove the last item - if you don't apply (a) first then (b)
will refer to an index off the end of the array.

Fixes #873