Raise Elected candidates to top of page
This makes them stand out a bit more and makes it easier to see at a
glance if there is a winner and who it is.

Do this by splitting current candidates into an elected and unelected
set and then iterating over those. Also adds a convenience has_elected
variable for checking if there are any elected candidates. If not then
we don't display the Elected Candidates section at all.

This also splits out the core of the candidate list display into a
partial as we use it in both the list of elected and unelected
candidates.

Fixes #692