Remove simple / complex fields before mirroring
By the time this is done, we'll have checked that the core database
tables are all empty, so this is highly likely to be the situation where
someone has just created a new database, run the migrations and is then
trying to import data from a live site.  In that situation, the
candidates_import_from_live_site command will create the simple and
complex fields from the remote site's API, so delete the fields that
were created by the initial migrations.