Add a config option to control hoisting elected candidates or not
When candidates are elected, we show them in a special section at the
top. People have different views about whether they should also be
included in the list of candidates below, however - this commit adds a
config option HOIST_ELECTED_CANDIDATES - if it's set to true, then
then elected candidates will only be shown in the 'elected' section.