Output all candidacies for people in the all candidates CSV file
Previously if a candidate was in multiple elections, then they would
only appear once in the "all candidates" CSV file.  This change makes
sure that they will appear once per candidacy. (It also deals with
the case that in an election-specific file, someone's standing for
more than one post.)

This also adds 'election_current' and 'election_date' columns to the
CSV output.

Fixes #891