Display candidate elected button on party list page
This is required in cases where the number of people elected from a
party list exceeds the number that are shown on the constituency page.
Without this the only way to mark people not shown on the constituency
page was to increase the party list length display limit for the
constituency page which is not something most users will be able to do.