Add the custom Modernizr build that Foundation recommends
Foundation's suggested HTML for 4.3.2 includes loading a custom build of
Modernizr, but we weren't including that in Javascript.  This meant that
the class="no-js" on the <html> element was never changed to "js"; also
it's probable that some of the Foundation Javascript also relied on
Modernizr so it seems best to include it.

I would normally include a non-minified version of a third party
library, since django-pipeline will compress it for production anyway,
but I can't see an obvious way of getting this page:

   http://modernizr.com/download/#-inlinesvg-svg-svgclippaths-touch-shiv-mq-cssclasses-teststyles-prefixes-ie8compat-load

... to generate a non-minified version :(

Modernizr is MIT-licensed: https://modernizr.com/license/