Allow election AreaType to be blank in forms/admin
Even though AreaType is a ManyToMany relationship and therefore option
unless blank=True is set on the field then the admin interface will
insist that you associate an election with an AreaType. We might not
want to do this if we are not using AreaType lookups, e.g. a simple YNR
instance which has no geolocation area lookup.

Fixes #942