Add a "wizard" view for adding a new candidacy for a person
When you have many elections, the person edit page was taking a huge
amount of time to load because it included form fields for every
election regardless of whether the person was marked as standing in
that election or not.  In order to avoid that being required, this
commit introduces a wizard for adding a candidacy for a person.
This view doesn't require any Javascript, so should be a good basis
for progressive enhancement.