Don't include candidacy fields for all elections on person edit page
With hundreds of elections this resulted in a huge amount of HTML being
generated for the candidacy forms, which could crash the browser in
trying to render it, and made the generation of the page very slow.

This commit removes the code that adds editable candidacy fields for
every election to the person edit page. A subsequent commit will replace
that with a link to the new "add candidacy" wizard.