Add a function + management command to check paired models
For all the *Extra models that are linked by a OneToOneField to a base
object, like PostExtra <-> Post, AreaExtra <-> Area we should never
have an object from the base model without exactly one corresponding
object of the extra model.  This can happen through coding errors,
manual changes to the database, or changes made in the Django admin.

This commit adds a function which will check for any such base objects
of these models that are unpaired. This can be called at the end of
transactions that make complex changes to the data to ensure that it
hasn't created that situation. It also provides this check as a
management command, for one-off checks.