Allow filtering out non-current elections in group_and_order_elections
When group_and_order_elections is used in PostListView, the template is
filtering out current elections. Over time, there will be more non-current
elections than current, so it makes sense to filter these out in SQL
rather than in the template. This over halves the query time in the
UK install.