Use Organization.classification in Extra slugs
This is a bit of a hack, I'm sorry.

Because OrganizationExtra.slug is unique in the DB and we can't
rely on globally unique slugs for all organisation types, we have to
use the Organization.classification to prefix the slug.

That way slugs only have to be unique per organisation type.

Alongside the importer change, there is also a migration.

It does a bit of a clean up of classifications too, as sometimes
they're missing.