Add a check function for MembershipExtra & PostExtraElection inconsistency
At the moment it's possible for the post and election referred to by a
Membership / MembershipExtra to not have a corresponding
PostExtraElection, which causes various obscure errors. We should
prevent these objects from being created in the first place, but in the
mean time this adds a function to check for whether this problem exists.