Prevent creation of bad MembershipExtra objects in the first place
As well as detecting the problem of the MembershipExtra objects whose
combination of post + election isn't in any PostExtraElection, we should
prevent them from being created in the first place.

This commit does that by overriding the save() method of
MembershipExtra. (We can't just do it with form validation, because most
of these kinds of problems are created by forgetting steps in some
management command that manipulates the date.)