Order type codes to avoid test failures from MapIt lookups
We only fake the MapIt responses in tests for a particular order of the
areas types (WMC, LAC) but the order of them in the URLs used by the
tests is sometimes different depending on the test environment and
Python version.  So this commit sorts the types before forming a URL
based on them and changes the faked URLs to match.