Reduce queries in checking group permission
Previously on each request we did 9 queries to find the groups the user
was a member of. This is unnecessary.

In addition, the group membership checking threw an exception if one of
the Group objects did not exist - this can happen in tests, for example,
if the Group would be created in a migration for an application not
included in INSTALLED_APPS.