The post_in_election template tag should work without uk_results
This branch had changed the implementation of the post_in_election
template tag so that it failed when the uk_results application had not
been migrated (e.g. in the non-country-specific tests). This commit
bring back the generic behaviour as a fallback when uk_results is not
present in INSTALLED_APPS.