Fix the test of fetching organizations from the API
The API for organizations starts from a query on OrganizationExtra. When
this test was written, there was no OrganizationExtra for the House of
Commons in the example UK data. We fixed this missing OrganizationExtra
in 9d3d7287c98c2be76b8 but didn't update this corresponding test. This
commit does that.