Upgrade django-nose to get the --keepdb option passthrough
Django 1.8 adds a very useful --keepdb option for running tests, which
on a project like this with a long migration history, will enormously
speed up running single tests repeatedly (a few seconds rather than
minutes). This commit upgrades django-nose to version that recognizes
the --keepdb option and passes it on to Django.