Use 'q' not 'postcode' when search redirects to the postcode finder
The search parameter used in the form on the front page is 'q', not
'postcode', so it'll make the form processing simpler in the future if
we use 'q' instead of 'postcode' when redirecting back to the home page.
This is only seen briefly by the user anyway, unless they've put in a
postcode-like but non-existent postcode.