Add a function to check all awkward constraints
In some cases (e.g. before committing a big update to the database) you
want to run all the checks we have on common ways the data can be
broken. Rather than having to remember which functions you'd have to run
for this, it's easier to have one check_constraints() function that does
all of those checks.