Prevent creating a MembershipExtra that conflicts with not_standing
We're having trouble working out how we've got a situation where some
candidates have a MembershipExtra.election that's also in their
not_standing relation. Since we can't figure out how this is arising,
this will raise an exception if an MembershipExtra is being saved that
would create that situation, so we can get a backtrace for when that
happens.

This commit doesn't address the other way this could happen, of course,
which is that an election that's already associated with their candidacy
could be added to not_standing. We'll add separate checks for that.