Add checks for existing candidacies before adding to not_standing
This is the other half of the checks in the previous commit: at each
point in the code where an election might be added to the not_standing
relation of PersonExtra, we check that there is no existing candidacy
of that election - if there is, an exception is thrown. The reasoning
is as before: we don't what bad logic is causing this situation to
arise, so we should throw an exception when that situation would be
created so we can get a backtrace to help in fixing that error.