Split out the Twitter avatar queuing from the Twitter username updater
The command for checking whether Twitter screen names have changed
had had code added to it to also import Twitter avatars of candidates
into the photo moderation queue. This is problematic because the
latter task is something that is only run occasionally, whereas the
screen name checker should be run every day to spot any changes of
screen name around elections.

This commit makes the following changes to fix that problem and make
this code easier to work with:

 * It factors out and simplifies the code for getting screen name
   and user ID data from the Twitter API.
 * It adds a new command just for queuing images.
 * It adds tests for both of these commands, with 100% coverage
 * It fixes the case where no results are found from the Twitter API's
   user lookup endpoint.