Output any changes made by candidates_update_twitter_usernames
The distinction between what was output by default and what required
the verbose option was a confusing in this command. This is typically
run from cron, with results reported by email and without the verbose
option - I think it's best if this command outputs *any* change it
makes in that context.

The verbose option makes sure that there's some output for every
person considered.

Hopefully this change makes the output less surprising.