Refactor UK get_areas_from_* functions to use EveryElection
Every election returns elections for a postcode or coords, and
always uses the same IDs for elections, as that's where we import
from.

This change removes the calls to MaPit and replaces them with simpler
calls to EE. It also merges the two get_areas_from_* functions
in to a single get_areas function.