Set default pointer value to None in get_version_diff
Previously this was failing for example when trying to compare a dict with a
given key to a dict without that key. For example with:

[
    {
        name: foo,
        other_names [{
            note: "something",
            name: "Jane"
        }]
    },
    {
        name: foo,
        other_names [{
            name: "Jane"
        }]
    },
]

`get_version_diff` was looking for the `note` key in the 2nd dict and
raising a JsonPointerException when it wasn't found. Adding a default
fixes this.

There is a valid question as to _why_ the key wasn't there, but this code
should be able to deal with cases where it's not.