Look for parties where the register is 'none'
This looks like a bug in the EC's website. Minor parties in GB or NI aren't
getting returned when filtering on the GB or NI register alone, even though
they are listed as being in that register.

Adding 'none' here means all parties get returned and minor parties are added.