improve bulk add duplicate detection when middle names used
SOPN forms tend to include the full name of a candidate including
middles names which makes it likely that these will be used when bulk
adding. Because the default operator for haystack is AND if we search
for John Quincy Adams it will miss out John Adams and we end up with
duplicate candidates. So in cases with more than two names search for
`John Quincy Adams OR John Adams` to pick up existing candidates
without the middle name.

Fixes #787