improve name search when middle names are used
Because the default operator for haystack is AND if we search
for John Quincy Adams it will miss out John Adams and we end up with
duplicate candidates. So in cases with more than two names search for
`John Quincy Adams OR John Adams` to pick up existing candidates
without the middle name.