Ensure we don't to add a Twitter avatar to the moderation queue twice
This changes the candidates_add_twitter_images_to_queue command to be
much more conservative about when it will enqueue a Twitter avatar for
moderation. It used to be that we enqueued a Twitter avatar for anyone
who didn't already have a rejected image in the moderation queue. This
meant, however, that if you'd rejected a Twitter avatar image from the
moderation queue, or it was still awaiting moderation, this command
would still requeue the image. This means that we can't run the command
from a cron job every night, or it would keep adding already moderated
images to the queue.

A more accurate way to deal with this would be to record a hash of
the images which are added to the moderation queue, and only enqueue
an image that hasn't already been queued for that person based on its
hash. However, that's a more involved change that the one in this
commit, which should nonetheless allow us to run the command from
cron without a risk of enqueuing duplicates.

Note that this command will still enqueue a Twitter avatar image for a
person who has an image already associated with them, so long as they
have never had an image in the moderation queue. This seems OK to me -
the Twitter avatar might well be a better or more up-to-date image than
whatever manually uploaded one we have, and it should be up to the
moderator to decide which should be their primary image.