Add tests that I hoped would reproduce a data loss on merge issue
I added these tests in an attempt to reproduce the problem seen in #118.
These don't reproduce the problem, unfortunately. My hope is that since
that broken merge was done some time ago (and in particular before merging
happened in a transaction) this problem may have been fixed by one of the
changes to the codebase since then. These tests may add some value in
the future, however, either by adding more involved examples of merging,
or by providing tests we can tweak to reproduce the problem if it's
still present and we understand why it happened.

Alternatively, this commit could be useful for cherry-picking back to an
earlier version of the code to try to reproduce the problem.