The constituency page should 404 if the post & election aren't linked
The constituency detail view page has a URL of the form:

  /election/ELECTION-ID/post/POST-ID/ARBITRARY-SLUG

Previously, if the POST-ID referred to a Post that wasn't associated
with the Election with ID ELECTION-ID, you'd just see an empty page.
This should really 404 instead as a more helpful signal to a user (who
might have been just been editing the URL - that's how this came up).

Thanks to Andy Lulham (@andylolz) for spotting this.