Make the Twitter-related commands handle people in a predictable order
The tests of the two commands that use the Twitter API would randomly
fail because they relied on the order in which the commands handled
Person objects. This commit changes those commands to process people in
alphabetical order and adjusts the test expectations to match that
order. (The time it takes to order people by name is neglible compared
to the running time of those commands.)