Add an overridable function for install-specific merge behaviour
For example, the UK installation of this codebase needs to care about
preserving the number of votes a candidate got when merging two people.
This commit adds a new function to candidates.election_specific called
additional_merge_actions which can be overridden to do additional
merge actions. It should be a context manager which yields to do the
usual merge actions (sometimes, as in that UK example, you need to
save data before the merge and act on it afterwards).