Avoid a bad queryset union, which caused person updates to fail
This fixes #134.

The line:

  self.elections_with_fields  = self.elections_with_fields | election

... was doing a union of two querysets. The results of this are
non-obvious (or incorrect depending on how you look at it) - even though
evaluating self.elections_with_fields would return 1 Election, and
evaluating election would return 1 Election, evaluating the union of
those two querysets would return about 500 instances of the latter
Election. The queries being generated by each queryset were as follows:

The self.elections_with_fields query was:

    SELECT "elections_election".[...]
    FROM "elections_election"
    INNER JOIN "candidates_membershipextra"
       ON ( "elections_election"."id" = "candidates_membershipextra"."election_id" )
    INNER JOIN "popolo_membership"
       ON ( "candidates_membershipextra"."base_id" = "popolo_membership"."id" )
    WHERE ("elections_election"."current" = True AND "popolo_membership"."person_id" = 5664)
    ORDER BY "elections_election"."election_date" DESC

The election query was:

    SELECT "elections_election".[...]
    FROM "elections_election"
    WHERE "elections_election"."slug" = parl.2017-06-08'

But the query generated for the union of those two querysets is:

    SELECT "elections_election".[...]
    FROM "elections_election"
    LEFT OUTER JOIN "candidates_membershipextra"
        ON ( "elections_election"."id" = "candidates_membershipextra"."election_id" )
    LEFT OUTER JOIN "popolo_membership"
        ON ( "candidates_membershipextra"."base_id" = "popolo_membership"."id" )
    WHERE (("elections_election"."current" = True AND "popolo_membership"."person_id" = 5664)
           OR "elections_election"."slug" = parl.2017-06-08)
    ORDER BY "elections_election"."election_date" DESC

In other words, it's including one Election with slug parl.2017-06-08
for every membership of that election, in addition to the elections the
candidate is standing in.

These many additional fields for the same election were processed after
the standing_ field for that election had been pop()ped which triggered
the issue in bug #134.

This is easily avoided, since there is not reason not to evaluated these
querysets immediately. This commit introduces that change.